DEV - Gif

+++ trab_2/utils.py
@@ -136,9 +136,11 @@ def make_gif(dir='gif_images'):
     with imageio.get_writer(ROOT_DIR/'movie.gif', mode='I',duration=0.1) as writer:
         for filename in sorted(os.listdir(ROOT_DIR/dir)):
             print(filename)
-            image = imageio.imread(str((ROOT_DIR/dir/filename)))
-            writer.append_data(image)
+            if filename.endswith(".png"):
+                image = imageio.imread(str((ROOT_DIR/dir/filename)))
+                writer.append_data(image)
     for image in os.listdir(ROOT_DIR/dir):
-        os.remove(ROOT_DIR/dir/image)
+        if image.endswith(".png"):
+            os.remove(ROOT_DIR/dir/image)
 if __name__ == '__main__':
     make_gif()
